backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-genai
pandas
openpyxl
